qingxiu-backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
Django==3.1.1
django-cors-headers==3.5.0
django-redis==4.12.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
minio==6.0.0
pinyin==0.4.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
uWSGI==2.0.19.1
xlrd==1.2.0
minio==6.0.0
celery==4.4.7
mongoengine==0.20.0
django-rest-framework-mongoengine==3.4.1
pymongo==3.11.0
baidu-aip==2.2.18.0
PyMuPDF==1.17.6
docxtpl==0.11.0
openpyxl==3.0.6
aliyunsdkcore
aliyun-python-sdk-core-v3
pypinyin
image
